used conditional; only rendering the 'delete' button if the prop 'deleteAnimal' exists

+++ src/components/animal/AnimalCard.js
@@ -10,27 +10,50 @@ const AnimalCard = (props) => {
   } catch {
     pictureFileId = require(`./animal.svg`);
   }
-  return (
-    <div className="card">
-      <div className="card-content">
-        <picture>
-          <img src={pictureFileId} alt="My Dog" />
-        </picture>
-        <h3>Name: <span className="card-petname">
-          {firstLetterCase(props.animal.name)}
-        </span></h3>
-        <p>Breed: {props.animal.breed}</p>
-        <Link to={`/animals/${props.animal.id}`}>
-          <button>Details</button>
-        </Link>
-        <button type="button"
-          onClick={() => props.history.push(`/animals/${props.animal.id}/edit`)}>
-          Edit
+  if (props.deleteAnimal) {
+    return (
+      <div className="card">
+        <div className="card-content">
+          <picture>
+            <img src={pictureFileId} alt="My Dog" />
+          </picture>
+          <h3>Name: <span className="card-petname">
+            {firstLetterCase(props.animal.name)}
+          </span></h3>
+          <p>Breed: {props.animal.breed}</p>
+          <Link to={`/animals/${props.animal.id}`}>
+            <button>Details</button>
+          </Link>
+          <button type="button"
+            onClick={() => props.history.push(`/animals/${props.animal.id}/edit`)}>
+            Edit
         </button>
-        <button type="button" onClick={() => props.deleteAnimal(props.animal.id)}>Discharge Animal</button>
+          <button type="button" onClick={() => props.deleteAnimal(props.animal.id)}>Discharge Animal</button>
+        </div>
       </div>
-    </div>
-  );
+    );
+  } else {
+    return (
+      <div className="card">
+        <div className="card-content">
+          <picture>
+            <img src={pictureFileId} alt="My Dog" />
+          </picture>
+          <h3>Name: <span className="card-petname">
+            {firstLetterCase(props.animal.name)}
+          </span></h3>
+          <p>Breed: {props.animal.breed}</p>
+          <Link to={`/animals/${props.animal.id}`}>
+            <button>Details</button>
+          </Link>
+          <button type="button"
+            onClick={() => props.history.push(`/animals/${props.animal.id}/edit`)}>
+            Edit
+        </button>
+        </div>
+      </div>
+    );
+  }
 }
 
 export default AnimalCard;
